project/parkingArea/pyproject.toml
[project]
name = "parkingarea"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "ultralytics>=8.3.223",
]
